feat(ImageCorpper): update AspectRatio type to float (#456)
* feat: 更改 AspectRatio 数据类型为 float

* chore: bump version 9.0.1

+++ src/components/BootstrapBlazor.ImageCropper/ImageCropperOption.cs
@@ -35,7 +35,7 @@ public class ImageCropperOption
     /// </summary>
     /// <remarks>默认情况下，裁剪框具有自由比例。 设置为1, 裁剪区默认正方形</remarks>
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public int? AspectRatio { get; set; }
+    public float? AspectRatio { get; set; }
 
     /// <summary>
     /// Re-render the cropper when resizing the window. default tue
